chore: make some fixes per PR comments
