+++ packages/shared/lib/contexts/onboarding/actions/completeOnboardingProcess.ts
@@ -1,14 +1,14 @@
 import { get } from 'svelte/store'
 
 import { createNewAccount } from '@core/account'
-import { addNewProfile, loadPersistedProfileIntoActiveProfile, login, ProfileType } from '@core/profile'
+import { login } from '@core/profile'
 
-import { ProfileRecoveryType, ProfileSetupType } from '../enums'
-import { convertOnboardingProfileToPersistedProfile } from '../helpers'
-import { onboardingProfile, updateOnboardingProfile } from '../stores'
+import { ProfileSetupType } from '../enums'
+import { onboardingProfile } from '../stores'
+
+import { addOnboardingProfile } from './addOnboardingProfile'
 import { cleanupOnboarding } from './cleanupOnboarding'
-import { restoreBackupFromFile } from './restoreBackupFromFile'
-import { verifyAndStoreMnemonic } from './verifyAndStoreMnemonic'
+import { initialiseSecretManager } from './initialiseSecretManager'
 
 export async function completeOnboardingProcess(): Promise<void> {
     addOnboardingProfile()
@@ -17,40 +17,4 @@ export async function completeOnboardingProcess(): Promise<void> {
     const recoverAccounts = get(onboardingProfile)?.setupType !== ProfileSetupType.New
     void login(recoverAccounts)
     void cleanupOnboarding()
-}
-
-function addOnboardingProfile(): void {
-    const _onboardingProfile = get(onboardingProfile)
-    const newProfile = convertOnboardingProfileToPersistedProfile(_onboardingProfile)
-    addNewProfile(newProfile)
-    loadPersistedProfileIntoActiveProfile(_onboardingProfile?.id)
-}
-
-async function initialiseSecretManager(): Promise<void> {
-    const _onboardingProfile = get(onboardingProfile)
-
-    const profileType = _onboardingProfile?.type
-    const profileSetupType = _onboardingProfile?.setupType
-    const profileRecoveryType = _onboardingProfile?.recoveryType
-
-    if (profileSetupType === ProfileSetupType.New) {
-        if (profileType === ProfileType.Software) {
-            if (_onboardingProfile?.mnemonic) {
-                await verifyAndStoreMnemonic()
-                updateOnboardingProfile({ mnemonic: null })
-            }
-        }
-    } else if (profileSetupType === ProfileSetupType.Recovered) {
-        if (profileRecoveryType === ProfileRecoveryType.Mnemonic) {
-            if (_onboardingProfile?.mnemonic) {
-                await verifyAndStoreMnemonic()
-                updateOnboardingProfile({ mnemonic: null })
-            }
-        } else if (profileRecoveryType === ProfileRecoveryType.Stronghold) {
-            await restoreBackupFromFile(
-                _onboardingProfile?.importFile as Buffer,
-                _onboardingProfile?.strongholdPassword
-            )
-        }
-    }
 }
