+++ packages/shared/lib/contexts/onboarding/actions/index.ts
@@ -1,3 +1,4 @@
+export * from './addOnboardingProfile'
 export * from './backupInitialStronghold'
 export * from './cleanupOnboardingProfileManager'
 export * from './cleanupOnboarding'
@@ -9,5 +10,6 @@ export * from './destroyIotaProfileManager'
 export * from './initialiseOnboardingProfile'
 export * from './initialisePincodeManager'
 export * from './initialiseProfileManagerFromOnboardingProfile'
+export * from './initialiseSecretManager'
 export * from './restoreBackupFromFile'
 export * from './verifyAndStoreMnemonic'
