+++ packages/shared/lib/contexts/onboarding/actions/initialiseOnboardingProfile.ts
@@ -1,6 +1,5 @@
 import { get } from 'svelte/store'
 
-import { AppStage, appStage } from '@core/app'
 import { profileManager } from '@core/profile-manager'
 
 import { buildOnboardingProfile } from '../helpers'
@@ -9,13 +8,12 @@ import { onboardingProfile } from '../stores'
 /**
  * Builds a new onboarding profile and sets the Svelte store accordingly.
  */
-export function initialiseOnboardingProfile(): Promise<void> {
+export function initialiseOnboardingProfile(isDeveloperProfile: boolean): Promise<void> {
     if (get(profileManager)) {
         console.error('Profile is already created')
         return
     }
 
-    const isDeveloperProfile = get(appStage) !== AppStage.PROD
     const _newProfile = buildOnboardingProfile(isDeveloperProfile)
     onboardingProfile.set(_newProfile)
 }
