+++ packages/shared/lib/contexts/onboarding/helpers/index.ts
@@ -1,6 +1,8 @@
 export * from './buildOnboardingProfile'
+export * from './buildProfileManagerOptionsFromOnboardingProfile'
 export * from './convertOnboardingProfileToPersistedProfile'
 export * from './getIotaProfileManagerStorageDirectory'
 export * from './getProfileTypeFromProfileRecoveryType'
 export * from './setProfileRecoveryTypeFromFilename'
+export * from './shouldUseDeveloperProfile'
 export * from './validateBackupFile'
