+++ packages/shared/lib/contexts/onboarding/helpers/shouldUseDeveloperProfile.ts
@@ -1,0 +1,7 @@
+import { get } from 'svelte/store'
+
+import { AppStage, appStage } from '@core/app'
+
+export function shouldUseDeveloperProfile(): boolean {
+    return get(appStage) !== AppStage.PROD
+}
