+++ packages/shared/lib/contexts/onboarding/types/index.ts
@@ -1,1 +1,2 @@
 export * from './import-file.type'
+export * from './profile-manager-options.type'
