+++ packages/shared/lib/contexts/onboarding/types/profile-manager-options.type.ts
@@ -1,0 +1,3 @@
+import { AccountManagerOptions } from '@iota/wallet'
+
+export type ProfileManagerOptions = AccountManagerOptions
